savva-tools/requirements.txt
qrcode[pil]
pymupdf
pillow
